+++ src/shared/components/event-management/components/no-tasks-available/no-tasks-available.component.spec.ts
@@ -1,0 +1,30 @@
+import { DebugElement } from '@angular/core';
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { By } from '@angular/platform-browser';
+import { RouterTestingModule } from '@angular/router/testing';
+import { NoTasksAvailableComponent } from './no-tasks-available.component';
+
+describe('NoTasksAvalaibleComponent', () => {
+  let component: NoTasksAvailableComponent;
+  let fixture: ComponentFixture<NoTasksAvailableComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      imports: [RouterTestingModule],
+      declarations: [NoTasksAvailableComponent]
+    })
+      .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(NoTasksAvailableComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should display error message no tasks available', () => {
+    const heading: DebugElement = fixture.debugElement.query(By.css('.govuk-heading-m'))
+    const headingHtml = heading.nativeElement as HTMLElement;
+    expect(headingHtml.innerText).toBe('No task available');
+  });
+});