+++ src/shared/components/event-management/components/no-tasks-available/no-tasks-available.component.ts
@@ -1,0 +1,9 @@
+import { Component, Input } from '@angular/core';
+
+@Component({
+  selector: 'app-no-tasks-available',
+  templateUrl: './no-tasks-available.component.html'
+})
+export class NoTasksAvailableComponent {
+  @Input() public caseId: string;
+}