+++ src/shared/components/event-management/components/task-assigned/task-assigned.component.spec.ts
@@ -1,0 +1,30 @@
+import { DebugElement } from '@angular/core';
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { By } from '@angular/platform-browser';
+import { RouterTestingModule } from '@angular/router/testing';
+import { TaskAssignedComponent } from './task-assigned.component';
+
+describe('TaskRequirementComponent', () => {
+  let component: TaskAssignedComponent;
+  let fixture: ComponentFixture<TaskAssignedComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      imports: [RouterTestingModule],
+      declarations: [TaskAssignedComponent]
+    })
+      .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(TaskAssignedComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should display error message task assigned', () => {
+    const heading: DebugElement = fixture.debugElement.query(By.css('.govuk-heading-m'))
+    const headingHtml = heading.nativeElement as HTMLElement;
+    expect(headingHtml.innerText).toBe('Task assignment required');
+  });
+});