+++ src/shared/components/event-management/models/index.ts
@@ -1,0 +1,2 @@
+export * from './event-states-enum.model';
+export * from './state-machine-states-enum.model';